refactor: Organise files into packages
