+++ .gitignore
@@ -3,6 +3,9 @@
 /out/
 /*.iml
 
+# vscode files
+.vscode/
+
 # Gradle build files
 /.gradle/
 /build/
