+++ src/main/java/Duke.java
@@ -1,3 +1,5 @@
+import bot.Bot;
+
 public class Duke {
     public static void main(String[] args) {
       // Initialize and start bot
