+++ src/main/java/bot/CommandType.java
@@ -1,5 +1,17 @@
+package bot;
+
 import java.util.HashMap;
 import java.util.Map;
+
+import commands.ByeCommand;
+import commands.Command;
+import commands.DeadlineCommand;
+import commands.DeleteCommand;
+import commands.DoneCommand;
+import commands.EmptyCommand;
+import commands.EventCommand;
+import commands.ListCommand;
+import commands.TodoCommand;
 
 public enum CommandType {
