+++ src/main/java/bot/InputManager.java
@@ -1,3 +1,5 @@
+package bot;
+
 import java.util.Scanner;
 
 /**
