+++ src/main/java/commands/DeleteCommand.java
@@ -1,4 +1,10 @@
-public class DoneCommand implements Command {
+package commands;
+
+import bot.Bot;
+import bot.OutputManager;
+import exceptions.InvalidTaskException;
+
+public class DeleteCommand implements Command {
 
   @Override
   public void run(Bot bot, String[] args) {
@@ -6,12 +12,7 @@ public class DoneCommand implements Command {
     if (index < 0 || index >= bot.getTaskList().size()) {
       throw new InvalidTaskException(OutputManager.ERROR_SIGNATURE + "This task does not exist in the task list!");
     }
-    Task task = bot.getTaskAt(index);
-    task.markDone();
-    bot.printOutput(new String[] {
-      "Nice! I've marked this task as done:",
-      OutputManager.TEXT_BLOCK_MARGIN + task.toString()
-    });
+    bot.removeTask(index);
   }
   
 }
