+++ src/main/java/exceptions/InvalidArgumentsException.java
@@ -1,5 +1,9 @@
+package exceptions;
+
 public class InvalidArgumentsException extends RuntimeException {
+
   public InvalidArgumentsException(String errorMessage) {
     super(errorMessage);
   }
+
 }
