+++ src/main/java/exceptions/InvalidTaskException.java
@@ -1,5 +1,9 @@
+package exceptions;
+
 public class InvalidTaskException extends RuntimeException {
+
   public InvalidTaskException(String errorMessage) {
     super(errorMessage);
   }
+
 }
