+++ src/main/java/tasks/DeadlineTask.java
@@ -1,20 +1,23 @@
-public class EventTask extends Task {
+package tasks;
+
+public class DeadlineTask extends Task {
+
   private String taskText;
   private String taskDetail;
 
-  public EventTask(String taskText, String taskDetail) {
+  public DeadlineTask(String taskText, String taskDetail) {
     this.taskText = taskText.trim();
     this.taskDetail = taskDetail.trim();
   }
 
   @Override
   String getTaskDesc() {
-    return String.format("%s (at: %s)", taskText, taskDetail);
+    return String.format("%s (by: %s)", taskText, taskDetail);
   }
 
   @Override
   String getTaskSymbol() {
-    return "E";
+    return "D";
   }
   
 }
