+++ src/main/java/tasks/EventTask.java
@@ -1,21 +1,23 @@
-public class DeadlineTask extends Task {
+package tasks;
+
+public class EventTask extends Task {
 
   private String taskText;
   private String taskDetail;
 
-  public DeadlineTask(String taskText, String taskDetail) {
+  public EventTask(String taskText, String taskDetail) {
     this.taskText = taskText.trim();
     this.taskDetail = taskDetail.trim();
   }
 
   @Override
   String getTaskDesc() {
-    return String.format("%s (by: %s)", taskText, taskDetail);
+    return String.format("%s (at: %s)", taskText, taskDetail);
   }
 
   @Override
   String getTaskSymbol() {
-    return "D";
+    return "E";
   }
   
 }
