+++ src/main/java/tasks/GeneralTask.java
@@ -1,3 +1,5 @@
+package tasks;
+
 public class GeneralTask extends Task {
 
   private String description;
